ux(lens): tie the waterfall's tooltip to the clamp that earns it
Review follow-ups on the pointer-vocabulary pass.

The column name carried its native `title` unconditionally, which is the very
thing the first commit here argues against: the tooltip is the full name behind
a two-line clamp, so on a name that fits it is a second copy of what is already
on screen. Whether there is a clamp is a question about the rendered box — the
same name fits at 1400px and is cut at 900px — and the runtime already had the
measurement for it, in a hook the metric strip uses to decide whether a clipped
caption joins its note. It moves next to the panels that ask it.

Two tests tightened with it: the aria-hidden sweep now counts a `title` on the
hidden element itself and not only on its descendants, and the composition blur
is asserted down to the leader line, which rides with the label it points from.
